Updated deceased datetime in FHIR exporter
Updated FHIR exporter to handle the rare case when a patient's death date exists as an event but still occurs in the future (the patient hasn't actually died yet). This is typically seen for patients with lung cancer and is a rare occurence (1-3 in 10k records). Also added the deceasedBoolean field to the FHIR records that are exported to make it easier to query FHIR for living or deceased patients using the Patient?deceased=true/false query.