Fixed opioid medications, consolidated incidence
The injuries module no longer ends opioid prescriptions - it's now up to the opioid addiction module to handle that. Put incidences in the same module as the injury pathways to make things simpler. Ran the tests and generated 100 patients. Nothing blew up, but the number of broken bones may be a little high. Going to generate 10k patients now and check the final prevalence stats.