Updated Pregnancy Module
Updated the pregnancy and birth modules (the birth module is now part of the pregnancy module) to include procedures, encounters, conditions, and careplans. For simplicity I removed the rare case of home birth or birth-before-arrival. Miscarriages may occur during the first or second trimester. Induced abortions always come after the 10-week visit, even though in reality a mother may not visit the doctor before electing to have an abortion.