Adjusted addiction and misuse instances
The prevalence of opioid addicts is now realistic - about 1% of the population. The updated injuries module now drives about 75% of the opioid prescriptions. The remainder are still generated by the opioids module. In general I reduced the incidence of misuse and changed the time scale of the module to 1-week steps instead of 1-day. I adjusted the relevant probabilities accordingly.