Added a reason to anything reasonable
Added reasons to all Encounter, Procedure, MedicationOrder, and CarePlanStart states that had a preceeding ConditionOnset. Also reordered a couple of things in some of the states for consistency. For example, transitions are now ALWAYS the last attribute in a state. Remarks are now ALWAYS lists, for consistency. I plan to issue a separate PR after a run through with a JSON formatter so all of our GMF modules have a consistent look and feel.