Fix FHIR Encounter class codes
When exporting Encounter FHIR resources, Synthea is currently outputting
a lowercased version of the EncounterType enum for the Encounter.class
element. However, this value should be from the ActEncounterCode VS:
http://www.hl7.org/implement/standards/fhir/v3/ActEncounterCode/vs.html

This change adds the appropriate value into the EncounterType enum and
updates the STU3 and R4 exporters to use it.