Changes to weight loss in patients under 20
Previously, the threshold for patients to start losing weight was a BMI
over 30. This doesn't make sense in a pediatric population and at
younger ages, hitting a BMI over 30 is close to physically impossible.

This adds the BMI percentiles from the CDC growth charts. It also
changes the threshold for weight management for patients under 20 to be
anyone at or above the 85th percentile for BMI for their age in months.