WIP: US Census definition of ethnicity
This change starts to move the code to using the US Census definition of
ethnicity, which is "Hispanic, Latino or Spanish" or "non-Hispanic".

The change decouples the selection of race and ethnicity, as the way
demographic information is currently kept doesn't really allow for that.

Language selection is now based on race and ethnicity. Probabilities
are now national, as opposed to based on MA (and probably weighted
incorrectly).

Foreign birth place is now based on language.

Several look up files have modifications to handle the changes or fix
previously missing data.

This code needs more testing and style fixes.