Closing ED Encounter for appendicitis
The appendicitis module was written that encounters were auto closing.
Since the ED encounter is never closed, in some cases it can remain open
for an incredibly long time (decades). This fixes it for the
appendicitis module, but there is still at least one case of a hanging
open encounter in the injury module.