Sometimes, we can delay death...
If an encounter has a procedure with a duration, the length of the
procedure will be added to the time for the encounter. If this is
immediately followed by a Death state, the death will happen before the
end of the encounter.

This code pushes death back to the time of the end of the encounter.
Death is inevitable, this code just delays it.