Adding procedure codes to Imaging Studies
Synthea ImagingStudy states can have procedure codes, but nothing really
happens with them in the simulation or export.

This takes the procedure code and adds it to the entry for ImagingStudy
in the HealthRecord. It then updates the export for FHIR STU3, R4 and
CSV so that the procedure code is exported. FHIR DSTU2 doesn't appear to
have a place in the ImagingStudy resource for procedure code.