Fix for Delays that could possibly have a negative time
Now, when a Delayable generates the length of delay, if it would be
negative, it's just set to zero. Essentially the Delay state will
complete at the current time.

Added tests to validate. Also added a test to ensure that Gaussian
delays actually generating values with a distribution mean close
to the specified mean.