Add a JavaScript engine for graphviz
By default, the graphviz-java library will use a local installation of
graphviz if available. If not, it will try to use a JavaScript based
version of graphviz. Since Java 15, the JDK no longer ships with a
JavaScript runtime, so this adds one to allow tests to pass on machines
running Java 15 or later without graphviz installed.

This should address issue #841