Fixes for employment condition and stress level
Previously, the SDoH HSRN submodule would change employment condition
every wellness visit, even if the actual condition did not change.

This change keeps the same condition across wellness visits of the
status does not change.

There was a similar problem with stress level. That is also fixed by
this change.