Fix issues with NPEs related to the employment module
The simulation expects every person to have an employment model set in
their attributes. If a person was generated an a county where the
`sdoh.csv` file did not have a row, the employment model would not get
set. This change uses the state average information when the desired
county can not be found.